Use .gitkeep to preserve directory structure; replace embed.go with embed_static.go
